+++ templates/part.content.php
@@ -40,17 +40,7 @@ style(
 	<div id='breadcrumbs'></div>
 	<div class="left">
 		<!-- sorting buttons -->
-		<div id="sort-date-button" class="button sorting right-switch-button">
-			<div class="flipper">
-				<img class="svg asc front" src="<?php print_unescaped(
-					image_path($_['appName'], 'dateasc.svg')
-				); ?>" alt="<?php p($l->t('Sort by date')); ?>"/>
-				<img class="svg des back" src="<?php print_unescaped(
-					image_path($_['appName'], 'datedes.svg')
-				); ?>" alt="<?php p($l->t('Sort by date')); ?>"/>
-			</div>
-		</div>
-		<div id="sort-name-button" class="button sorting left-switch-button">
+		<div id="sort-name-button" class="button sorting" title="<?php p($l->t('Sort by name')); ?>">
 			<div class="flipper">
 				<img class="svg asc front" src="<?php print_unescaped(
 					image_path($_['appName'], 'nameasc.svg')
@@ -58,6 +48,26 @@ style(
 				<img class="svg des back" src="<?php print_unescaped(
 					image_path($_['appName'], 'namedes.svg')
 				); ?>" alt="<?php p($l->t('Sort by name')); ?>"/>
+			</div>
+		</div>
+		<div id="sort-date-button" class="button sorting" title="<?php p($l->t('Sort by modified date')); ?>">
+			<div class="flipper">
+				<img class="svg asc front" src="<?php print_unescaped(
+					image_path($_['appName'], 'dateasc.svg')
+				); ?>" alt="<?php p($l->t('Sort by modified date')); ?>"/>
+				<img class="svg des back" src="<?php print_unescaped(
+					image_path($_['appName'], 'datedes.svg')
+				); ?>" alt="<?php p($l->t('Sort by modified date')); ?>"/>
+			</div>
+		</div>
+		<div id="sort-date-taken-button" class="button sorting" title="<?php p($l->t('Sort by EXIF taken date')); ?>">
+			<div class="flipper">
+				<img class="svg asc front" src="<?php print_unescaped(
+					image_path($_['appName'], 'takenasc.svg')
+				); ?>" alt="<?php p($l->t('Sort by EXIF taken date')); ?>"/>
+				<img class="svg des back" src="<?php print_unescaped(
+					image_path($_['appName'], 'takendes.svg')
+				); ?>" alt="<?php p($l->t('Sort by EXIF taken date')); ?>"/>
 			</div>
 		</div>
 	</div>
@@ -74,16 +84,16 @@ style(
 	<div id="file_action_panel"></div>
 	<span class="right">
 		<!-- sharing button -->
-		<div id="shared-button" class="button">
+		<div id="shared-button" class="button" title="<?php p($l->t('Share')); ?>">
 			<img class="svg" src="<?php print_unescaped(
 				image_path('core', 'actions/shared.svg')
-			); ?>" alt="<?php p($l->t("Share")); ?>"/>
+			); ?>" alt="<?php p($l->t('Share')); ?>"/>
 		</div>
 		<a class="share" data-item-type="folder" data-item=""
-		   title="<?php p($l->t("Share")); ?>"
+		   title="<?php p($l->t('Share')); ?>"
 		   data-possible-permissions="31"></a>
 		<!-- info button -->
-		<div id="album-info-button" class="button">
+		<div id="album-info-button" class="button" title="<?php p($l->t('Album information')); ?>">
 			<span class="ribbon black"></span>
 			<img class="svg" src="<?php print_unescaped(
 				image_path('core', 'actions/info.svg')
@@ -94,7 +104,7 @@ style(
 			<div class="album-info-content markdown-body"></div>
 		</div>
 		<!-- button for opening the current album as file list -->
-		<div id="filelist-button" class="button view-switcher gallery">
+		<div id="filelist-button" class="button view-switcher gallery" title="<?php p($l->t('File list')); ?>">
 			<div id="button-loading" class="hidden"></div>
 			<img class="svg" src="<?php print_unescaped(
 				image_path('core', 'actions/toggle-filelist.svg')